Move graphics to separate file
